Move `integration-tests` outside of consumers module
