--- activity/src/main/java/dev/marcellogalhardo/retained/activity/test/EmptyActivity.kt
@@ -1,5 +1,0 @@
-package dev.marcellogalhardo.retained.activity.test
-
-import androidx.activity.ComponentActivity
-
-internal class EmptyActivity : ComponentActivity()
