+++ core/src/main/AndroidManifest.xml
@@ -1,7 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.marcellogalhardo.retained.core">
-
-    <application>
-        <activity android:name="dev.marcellogalhardo.retained.core.dev.marcellogalhardo.retained.activity.test.EmptyActivity" />
-    </application>
-</manifest>
+<manifest package="dev.marcellogalhardo.retained.core" />
