+++ integration-tests/build.gradle
@@ -1,0 +1,68 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+
+android {
+    compileSdkVersion 30
+    buildToolsVersion "30.0.2"
+
+    defaultConfig {
+        minSdkVersion 21
+        targetSdkVersion 30
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles 'consumer-rules.pro'
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_1_8.toString()
+        freeCompilerArgs += ["-Xopt-in=kotlin.RequiresOptIn", "-Xexplicit-api=strict"]
+    }
+
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+        }
+    }
+}
+
+dependencies {
+    implementation projects.core
+    implementation projects.activity
+    implementation projects.fragment
+
+    implementation libs.kotlin.stdlib
+
+    implementation libs.androidx.viewmodel.core
+    implementation libs.androidx.viewmodel.savedstate
+
+    implementation libs.androidx.activity.core
+    implementation libs.androidx.fragment.core
+
+    // Unit Testing
+    testImplementation libs.junit.core
+    testImplementation libs.robolectric.core
+    testImplementation libs.androidx.test.core
+    testImplementation libs.androidx.test.runner
+    testImplementation libs.androidx.test.rules
+    testImplementation libs.androidx.test.junit
+    testImplementation libs.androidx.test.truth
+    testImplementation libs.google.truth
+
+    // Android Testing
+    debugImplementation libs.androidx.fragment.testing
+    androidTestImplementation libs.androidx.test.junit
+}
