+++ integration-tests/src/main/java/dev/marcellogalhardo/retained/activity/EmptyActivity.kt
@@ -1,0 +1,5 @@
+package dev.marcellogalhardo.retained.activity
+
+import androidx.activity.ComponentActivity
+
+internal class EmptyActivity : ComponentActivity()
